First Upload
First upload of the site. I'd liked to change the way I did the pages. Right now they scroll up but i think it'd be better with a fadeIn maybe? With them all stacked on top of each other. The recent blogs solutions on the front page is working well. However there is a slight delay when you manually close the nav on pages, I'm not sure why yet.